Split error/exception handling
